+++ server/impl/src/main/java/com/walmartlabs/concord/server/cfg/ProcessStateConfiguration.java
@@ -52,6 +52,10 @@ public class ProcessStateConfiguration implements Serializable {
     private boolean eventsCleanup;
 
     @Inject
+    @Config("process.eventStatsCleanup")
+    private boolean eventStatsCleanup;
+
+    @Inject
     @Config("process.logsCleanup")
     private boolean logsCleanup;
 
@@ -62,7 +66,6 @@ public class ProcessStateConfiguration implements Serializable {
     @Inject
     @Config("process.maxStateAge")
     private long maxStateAge;
-
     @Inject
     @Config("process.secureFiles")
     private List<String> secureFiles;
@@ -103,6 +106,10 @@ public class ProcessStateConfiguration implements Serializable {
 
     public boolean isEventsCleanup() {
         return eventsCleanup;
+    }
+
+    public boolean isEventStatsCleanup() {
+        return eventStatsCleanup;
     }
 
     public boolean isLogsCleanup() {